+++ app/controllers/landable/api/template_audits_controller.rb
@@ -16,9 +16,10 @@ module Landable
         respond_with audit
       end
 
-      def audit_params
-        params.require(:template_audit).permit(:id, :approver, :notes, flags: [])
-      end
+      private
+        def audit_params
+          params.require(:template_audit).permit(:id, :approver, :notes, flags: [])
+        end
     end
   end
 end