+++ app/models/landable/page_revision.rb
@@ -10,6 +10,7 @@ module Landable
       'page_id',
       'imported_at',
       'created_at',
+      'deleted_at',
       'updated_at',
       'published_revision_id',
       'is_publishable',
@@ -22,7 +23,8 @@ module Landable
 
     belongs_to :author
     belongs_to :page, inverse_of: :revisions
-    has_many   :screenshots, class_name: 'Landable::Screenshot', as: :screenshotable
+
+    after_commit :add_screenshot!, on: :create
 
     def page_id=(id)
       # set the value
@@ -43,14 +45,14 @@ module Landable
     end
 
     def snapshot
-      Page.new(title: self.title, 
-               meta_tags: page.meta_tags, 
+      Page.new(title: self.title,
+               meta_tags: page.meta_tags,
                head_content: page.head_content,
-               body: self.body, 
-               path: self.path, 
-               redirect_url: self.redirect_url, 
-               status_code: self.status_code, 
-               theme_id: self.theme_id, 
+               body: self.body,
+               path: self.path,
+               redirect_url: self.redirect_url,
+               status_code: self.status_code,
+               theme_id: self.theme_id,
                category_id: self.category_id,
                abstract: self.abstract,
                hero_asset_id: self.hero_asset_id)
@@ -85,11 +87,52 @@ module Landable
     end
 
     def preview_url
-      public_preview_page_revision_url(self)
+      begin
+        public_preview_page_revision_url(self, host: Landable.configuration.public_host)
+      rescue ArgumentError
+        Rails.logger.warn "Failed to generate preview url for page revision #{id} - missing Landable.configuration.public_host"
+        nil
+      end
     end
 
     def preview_path
       public_preview_page_revision_path(self)
     end
+
+    mount_uploader :screenshot, Landable::AssetUploader
+
+    def screenshot_url
+      screenshot.try(:url)
+    end
+
+    def add_screenshot!
+      return nil if preview_url.blank?
+
+      unless Landable.configuration.screenshots_enabled
+        Rails.logger.info "Screenshots disabled; skipping for #{path}"
+        return
+      end
+
+      attempts_left = 3
+
+      begin
+        attempts_left -= 1
+
+        self.screenshot = ScreenshotService.capture(preview_url)
+
+        # we've got a trigger preventing updates to other columns, so! muck
+        # about under the hood to commit the asset, and explicitly only update
+        # this column.
+        store_screenshot!
+        write_screenshot_identifier
+        update_column :screenshot, self[:screenshot]
+
+      rescue ScreenshotService::Error => error
+        Rails.logger.warn "Failed to generate screenshot (#{attempts_left} attempt(s) left) for #{path}: #{error.inspect}"
+
+        retry if attempts_left > 0
+      end
+    end
   end
+
 end