+++ app/models/landable/template.rb
@@ -1,6 +1,10 @@
 module Landable
   class Template < ActiveRecord::Base
     include Landable::TableName
+    include Landable::Librarian
+
+    # attributes
+    attr_accessor :temp_author
 
     validates_presence_of   :name, :slug, :description
     validates_uniqueness_of :name, case_sensitive: false
@@ -16,6 +20,12 @@ module Landable
     before_save -> template {
       template.is_publishable = true unless template.published_revision_id_changed?
     }
+
+    def deactivate
+      publish!(author_id: temp_author.id, notes: "This template has been trashed")
+
+      super
+    end
 
     def name= val
       self[:name] = val