+++ app/serializers/landable/template_serializer.rb
@@ -3,7 +3,7 @@ module Landable
     attributes :id, :name, :body, :description
     attributes :thumbnail_url, :slug, :is_layout, :editable
     attributes :file, :is_publishable
-    attributes :audit_flags
+    attributes :deleted_at, :audit_flags
 
     embed    :ids
     has_one  :published_revision