+++ app/serializers/landable/theme_serializer.rb
@@ -1,6 +1,6 @@
 module Landable
   class ThemeSerializer < ActiveModel::Serializer
-    attributes :id, :name, :editable, :body, :description, :thumbnail_url
+    attributes :id, :name, :editable, :body, :description, :thumbnail_url, :deleted_at
 
     embed :ids
   end