+++ db/migrate/20140501171359_add_deleted_at_to_assets.rb
@@ -1,0 +1,5 @@
+class AddDeletedAtToAssets < ActiveRecord::Migration
+  def change
+    add_column "#{Landable.configuration.database_schema_prefix}landable.assets", :deleted_at, :timestamp
+  end
+end