+++ db/migrate/20140501171406_add_deleted_at_to_templates.rb
@@ -1,0 +1,5 @@
+class AddDeletedAtToTemplates < ActiveRecord::Migration
+  def change
+    add_column "#{Landable.configuration.database_schema_prefix}landable.templates", :deleted_at, :timestamp
+  end
+end