+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 See README.md before updating this file.
 
-## Unreleased [#](https://github.com/enova/landable/compare/v1.7.1.rc1...master)
+## Unreleased [#](https://github.com/enova/landable/compare/v1.8.0...master)
+
+## 1.8.0 [#](https://github.com/enova/landable/compare/v1.7.1.rc1...v1.8.0)
+* Removing geminabox [#11]
+* Some fixes to the DB tests [#9]
 * Rails 4.x support [#8]
+* Screenshots by revision [#7]
+* Soft deletes [#6]
 
 ## 1.7.1.rc1 [#](https://github.com/enova/landable/compare/v1.7.0...v1.7.1.rc1)
 * Allow for revisions for Templates [#3]